requirements.txt
# Core
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
peft>=0.10.0
trl>=0.8.6
accelerate>=0.28.0

# Unsloth (CUDA 12.1 기준 — 환경에 맞게 수정)
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git

# Experiment tracking
mlflow>=2.12.0

# Config management
hydra-core>=1.3.0

# Merging
mergekit @ git+https://github.com/arcee-ai/mergekit.git

# Evaluation
evaluate>=0.4.0
rouge_score>=0.1.2

# Utilities
numpy>=1.24.0
tqdm>=4.66.0
